moved tree state to parent so it persists between renders
