+++ test-extension/src/containers/ParentTab.tsx
@@ -1,19 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import Tabs from '@mui/material/Tabs';
-import Tab from '@mui/material/Tab';
-import Box from '@mui/material/Box';
-import { ParentTabsProps } from '../types';
-import CustomTabPanel from '../components/CustomTabPanel';
-import a11yProps from '../functions/a11yProps';
-import TreeTab from './TreeTab';
-import QueriesTab from './QueriesTab';
+import React, { useEffect, useState } from "react";
+import Tabs from "@mui/material/Tabs";
+import Tab from "@mui/material/Tab";
+import Box from "@mui/material/Box";
+import { ParentTabsProps } from "../types";
+import CustomTabPanel from "../components/CustomTabPanel";
+import a11yProps from "../functions/a11yProps";
+import TreeTab from "./TreeTab";
+import QueriesTab from "./QueriesTab";
 
 const ParentTab = ({
   queryEvents,
   selectedQueries,
   handleSelectionChange,
   devToolsPort,
-  treeData
+  treeData,
 }: ParentTabsProps) => {
   const [value, setValue] = React.useState(0);
 
@@ -21,16 +21,35 @@ const ParentTab = ({
     setValue(newValue);
   };
 
+  //state for navigating between tree and charts
+  const [profilingStatus, setProfilingStatus] = useState<boolean>(false);
+
+  function sendMessageToContentScript(profilingEnabled: boolean) {
+    // only send message if devToolsPort is available and profiling is enabled
+    if (devToolsPort && profilingEnabled) {
+      devToolsPort.postMessage({
+        type: "profiling-status",
+        payload: profilingEnabled,
+      });
+    }
+  }
+  const toggleProfiling = () => {
+    console.log("toggleProfiling clicked");
+    const newProfilingStatus = !profilingStatus;
+    setProfilingStatus(newProfilingStatus);
+    sendMessageToContentScript(newProfilingStatus);
+  };
+
   return (
     <Box
       sx={{
-        width: '100%',
-        height: '100vh',
-        display: 'flex',
-        flexDirection: 'column',
+        width: "100%",
+        height: "100vh",
+        display: "flex",
+        flexDirection: "column",
       }}
     >
-      <Box sx={{ borderBottom: 1, borderColor: 'divider', height: '3rem' }}>
+      <Box sx={{ borderBottom: 1, borderColor: "divider", height: "3rem" }}>
         <Tabs
           value={value}
           onChange={handleChange}
@@ -42,7 +61,7 @@ const ParentTab = ({
       </Box>
 
       <Box
-        sx={{ flexGrow: 1, height: 'calc(100vh - 3rem)', paddingTop: '0.5rem' }}
+        sx={{ flexGrow: 1, height: "calc(100vh - 3rem)", paddingTop: "0.5rem" }}
       >
         <CustomTabPanel value={value} index={0}>
           <QueriesTab
@@ -56,7 +75,8 @@ const ParentTab = ({
         <CustomTabPanel value={value} index={1}>
           <TreeTab
             treeData={treeData}
-            devToolsPort={devToolsPort}
+            toggleProfiling={toggleProfiling}
+            profilingStatus={profilingStatus}
           />
         </CustomTabPanel>
       </Box>
